Fixed hole in assertArraysEqualIgnoringOrder where the assert passed incorrectly where the arrays contain repeated items, ie:
[item1, item1, item2] equals [item1, item2, item2]

These are now found to be not equal.

Added tests for this to the asunit3 and asunit4 suites (which were failing before the change to Assert), and updated the Assert in both suites so that they now pass.

Refactored logic for bothNull, eitherNull and differentLengths tests within asserts to use helper functions.  Refactored the null defense in assertEquals to give a cleaner early exit where expected is null.